Refactor OrderDetail.vue by extracting components for OrderHeader, OrderTimeline, OrderItemsList, and DeliveryInfo; removed redundant logic and improved readability.
